+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - new command `New-DataJourney`.
+- new command `Export-DataJourney`.
+- new command `Import-DataJourney`.
 - new command `New-DataLayer`.
 - new command `New-DataModel`.
 - new command `New-DataFlow`.